backend/requirements.txt
Flask
dnspython
flask_cors